auto vehicle ws_purchasePrice 형식 변경

+++ src/main/java/com/hyundaimotors/hmb/cdppapp/dto/IFHMBSAPEAICDPP0010/IFHMBSAPEAICDPP0010Dto.java
@@ -21,7 +21,7 @@ public class IFHMBSAPEAICDPP0010Dto {
         private String internalColorCode;
         private String externalColorCode;
         private String fuel;
-        private int purchasePrice;
+        private float purchasePrice;
         private String manufactureYear;
         private int modelYear;
         private int mileage;
